+++ odoo/local-src/connector_qoqa/qoqa_backend/common.py
@@ -64,12 +64,6 @@ class QoqaBackend(models.Model):
     token = fields.Char('Token')
     debug = fields.Boolean('Debug')
 
-    property_voucher_journal_id = fields.Many2one(
-        comodel_name='account.journal',
-        string='Journal for Vouchers',
-        domain="[('type', '=', 'general')]",
-        company_dependent=True,
-    )
     promo_type_ids = fields.One2many(
         comodel_name='qoqa.promo.type',
         inverse_name='backend_id',
@@ -331,12 +325,6 @@ class QoqaBackend(models.Model):
         return True
 
     @api.multi
-    def import_discount_accounting(self):
-        self._import_from_date('qoqa.discount.accounting',
-                               'import_discount_accounting_from_date')
-        return True
-
-    @api.multi
     def import_offer(self):
         self._import_from_date('qoqa.offer', 'import_offer_from_date')
         return True
@@ -377,12 +365,6 @@ class QoqaBackend(models.Model):
     @api.multi
     def import_one_offer(self):
         self._import_one('qoqa.offer', 'import_offer_id')
-        return True
-
-    @api.multi
-    def import_one_discount_accounting(self):
-        self._import_one('qoqa.discount.accounting',
-                         'import_discount_accounting_id')
         return True
 
     @api.multi